Let Xcode have its way upgrading project files. Move some JSON parsing into the model classes.

+++ iOS/Classes/SettingsViewController.m
@@ -51,7 +51,7 @@
 
 - (IBAction)degreesChanged:(id)sender
 {
-    int index = [self.degreesControl selectedSegmentIndex];
+    NSInteger index = [self.degreesControl selectedSegmentIndex];
     [[UtilityMethods sharedInstance] setCelsiusMode:(index == 1)];    
 }
